finance calculator final answer in integer

+++ python_introduction/finance_calculator.py
@@ -7,7 +7,7 @@ interest_rate = 0.05  # Assuming a 5% annual interest rate
 monthly_interest = monthly_savings * interest_rate
 
 
-projected_annual_savings = (monthly_savings + monthly_interest) * 12
+projected_annual_savings = int((monthly_savings + monthly_interest) * 12)
 
 print(f"Your monthly savings are: ${monthly_savings}")
 print(f"Projected savings after one year, with interest is : ${projected_annual_savings}")
